Some work on variable type consistancy, and getting rid of 64 bit processor checks.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@1214 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/zerospu2/misc.h
@@ -23,7 +23,7 @@
 
 #define Sleep(ms) usleep(1000*ms)
 
-inline unsigned long timeGetTime()
+inline u32 timeGetTime()
 {
 #ifdef _WIN32
 	_timeb t;
@@ -33,7 +33,7 @@ inline unsigned long timeGetTime()
 	ftime(&t);
 #endif
 
-	return (unsigned long)(t.time*1000+t.millitm);
+	return (u32)(t.time*1000+t.millitm);
 }
 
 #include <sys/time.h>
